% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BE_ADMIN_AGGLOMERATIONS}
\alias{BE_ADMIN_AGGLOMERATIONS}
\title{Maps with administrative boundaries of agglomerations (200m) of Belgium extracted from Open Data at Statistics Belgium}
\source{
\url{http://statbel.fgov.be/nl/binaries/SH_CENSUS_2011_AGGLOMERATIONS_200M.shp_tcm325-276234.zip}
}
\description{
Maps with administrative boundaries of agglomerations (200m) of Belgium extracted from Open Data at Statistics Belgium.
Agglomerations (200m) are groups of buildings of which none of them are more than 200m away from the closest neighbouring building.
}
\examples{
\dontrun{
data(BE_ADMIN_AGGLOMERATIONS)
library(sp)
plot(BE_ADMIN_AGGLOMERATIONS)
}
}
\references{
\url{http://statbel.fgov.be/nl/statistieken/opendata/datasets/tools/}
}

