% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BE_ADMIN_HIERARCHY}
\alias{BE_ADMIN_HIERARCHY}
\title{Definitions of sectors/municipalities/districts/provinces/regions in Belgium.}
\source{
\url{http://statbel.fgov.be/nl/binaries/TF_PSNL_INC_TAX_SECTOR_tcm325-278417.zip}
}
\description{
Definitions of sectors/municipalities/districts/provinces/regions in Belgium.
}
\examples{
data(BE_ADMIN_HIERARCHY)
str(BE_ADMIN_HIERARCHY)
}
\references{
\url{http://statbel.fgov.be/nl/statistieken/opendata/datasets/werk}
}

